refactor according to Google C++ style guide
