+++ src/preprocessor/elem_service/IElement.h
@@ -3,7 +3,7 @@
 #include <memory>
 #include <boost/numeric/ublas/matrix.hpp>
 #include <boost/numeric/ublas/vector.hpp>
-#include "../node/Node.h"
+#include "../node/node.h"
 
 using namespace boost::numeric;
